third_party/blink/web_tests/flag-specific/highdpi/fast/events/touch/gesture/gesture-scroll-by-page-expected.txt

This is a testharness.js-based test.
FAIL gesture-scroll-by-page assert_greater_than_equal: expected a number greater than or equal to 1050 but got 1049.3333740234375
Harness: the test ran to completion.
